tweaks
Changed the logic for BP_Door so it opens and closed based on if the player presses z(I will probably change this later when I add other interactable objects). Changed BP_TankCharacter movement to only move when pressing w or s, and rotate if they press a or d. Made the camera system in the west dinning hall work. But still need to tweak and play around with the trigger boxes and cameras